Updated main.dart to include modal Bottomsheet. I couldn't get it to run because of problems with my development environment. Someone else should try to see if they can run it.
If it runs, see if a sheet slides from the bottom of the screen when you tap a marker. If it doesn't, check the comments I've added to the code. There is an alternative way of adding the Bottomsheet by creating a new widget which I added as commented code at the bottom of this file. 

Also, I can collaborate with anyone on getting it implemented next Sprint if necessary.